Conflicts:
	src/main/java/org/goafabric/spring/boot/exampleservice/configuration/SecurityConfiguration.java

+++ src/main/resources/db/migration/V2__security.sql
@@ -1,7 +1,7 @@
 CREATE TABLE users (
   username VARCHAR(50) NOT NULL,
   password VARCHAR(100) NOT NULL,
-  enabled TINYINT NOT NULL DEFAULT 1,
+  enabled SMALLINT NOT NULL DEFAULT 1,
   PRIMARY KEY (username)
 );
